Create partial class CustomAction.Utils
