+++ src/Setup/DeviceHive.Setup.Actions/CustomAction.cs
@@ -13,7 +13,7 @@ using MongoDB.Driver;
 
 namespace DeviceHive.Setup.Actions
 {
-    public class CustomActions
+    public partial class CustomActions
     {
         #region Constants
 
@@ -188,14 +188,14 @@ namespace DeviceHive.Setup.Actions
 
                 int numRows = 0;
 
-                CreateRecordItem(certView, numRows++, "SSL_CERTIFICATE", "Not Selected", "Not Selected");
+                CreateComboBoxRecordItem(certView, numRows++, "SSL_CERTIFICATE", "Not Selected", "Not Selected");
 
                 foreach (var certificate in certificates)
                 {
                     session.Log("Subject: {0}; PublicKey: {1}", certificate.Subject, certificate.SubjectName.Name);
                     if (certificate.Subject.Trim() != String.Empty)
                     {
-                        CreateRecordItem(certView, numRows++, "SSL_CERTIFICATE", GetCertificateFriendlyName(certificate), certificate.Subject);
+                        CreateComboBoxRecordItem(certView, numRows++, "SSL_CERTIFICATE", GetCertificateFriendlyName(certificate), certificate.Subject);
                     }
                 }
             }
@@ -527,54 +527,6 @@ namespace DeviceHive.Setup.Actions
                 return value >= 1 && value < 65536;
             }
             return false;
-        }
-
-        private static bool GetPropertyBoolValue(Session session, string propertyName, bool required = false)
-        {
-            string propertyValue = session[propertyName];
-            if (string.IsNullOrEmpty(propertyValue))
-            {
-                string errorMessage = string.Format("{0} property has not been set up.", propertyName);
-                session.Log(errorMessage);
-                if (required)
-                {
-                    throw new ArgumentException(errorMessage);
-                }
-                return false;
-            }
-
-            bool result;
-            if (!bool.TryParse(propertyValue, out result))
-            {
-                throw new ArgumentException(string.Format("{0} property has wrong type value.", propertyName));
-            }
-            return result;
-        }
-
-        private static string GetPropertyStringValue(Session session, string propertyName, bool required = false)
-        {
-            string propertyValue = session[propertyName];
-            if (string.IsNullOrEmpty(propertyValue))
-            {
-                string errorMessage = string.Format("{0} property has not been set up.", propertyName);
-                session.Log(errorMessage);
-                if (required)
-                {
-                    throw new ArgumentException(errorMessage);
-                }
-                return string.Empty;
-            }
-            return propertyValue.Trim();
-        }
-
-        private static void CreateRecordItem(View view, int numRows, string propertyName, string text, string value)
-        {
-            Record record = new Record(4);
-            record.SetString(1, propertyName);
-            record.SetInteger(2, numRows);
-            record.SetString(3, value);
-            record.SetString(4, text);
-            view.Modify(ViewModifyMode.InsertTemporary, record);
         }
 
         private static string GetCertificateFriendlyName(X509Certificate2 certificate)
