+++ src/Setup/DeviceHive.Setup.Actions/DeviceHive.Setup.Actions.csproj
@@ -58,6 +58,7 @@
     <Compile Include="Credentials\CredentialManager.cs" />
     <Compile Include="Credentials\MongoDBCredentialManager.cs" />
     <Compile Include="Credentials\SqlCredentialManager.cs" />
+    <Compile Include="CustomAction.Utils.cs" />
     <Compile Include="Validation\AdministratorCredentialsValidator.cs">
       <SubType>Code</SubType>
     </Compile>
